Add ability to keep all working files
Currently only saving relevant files to disk. For debugging purposes,
it may be helpful to keep all files. Added a flag "--keep-tmp", which
will create a zip archive with the name "tmp_files-{timestamp}.zip" to
a separate "archives" folder in the output. This functionality extends
the save method.

Note: The archive function can also be used just to create an archive
instead of individual files if desired. However, this function is not
used in the current workflow.